requirements.txt
google-cloud-aiplatform>=1.40.0
google-cloud-firestore>=2.19.0
google-cloud-speech>=2.20.0
google-cloud-texttospeech>=2.14.0
google-auth>=2.23.0
google-genai>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
aiohttp>=3.9.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
itsdangerous>=2.1.0
razorpay>=1.4.0
stripe>=7.0.0
twilio>=8.0.0
python-pptx>=0.6.21
reportlab>=4.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
